Add defs, implement exePath on Haiku, remove redundant FileScanner code
Essentially everywhere __unix__ is checked for, __HAIKU__ is also now
checked for, because while being POSIX-compliant and essentially unix
enough for us, it's technically not unix.

exePath is now properly obtained on Haiku by doing the universally
agreed (see: HaikuPorts) way of doing so, which is looping through the
open images until we find one that matches us, getting it's "name",
and running dirname and realpath on the image name.

sys/sockio.h is now also included on Haiku in network enabled builds
as that's what actually contains the SIOC* macros.

Some redundant checking in FileScannerUnix's directory checking
was also removed, where it would check the node's d_type;
which is not in the POSIX standard, and not present on Haiku; and would
then see if it's equal to DT_DIR.

If that check failed, it would fallback to the POSIX-compliant method,
which was running a stat and S_ISDIR on it. This was deemed redundant,
as if the first check fails, we just do a second check, but a more
reliable and compatible check.

As such, the initial d_type check is now no longer present.